+++ modules/tinymce/src/plugins/table/main/ts/ui/Buttons.ts
@@ -239,6 +239,15 @@ const addButtons = (editor: Editor, selectionTargets: SelectionTargets, clipboar
     ),
     onSetup: selectionTargets.onSetupCellOrRow
   });
+
+  editor.ui.registry.addToggleButton('tablecaption', {
+    tooltip: 'Table caption',
+    onAction: () => {
+      editor.execCommand('mceTableToggleCaption');
+    },
+    icon: 'table-caption',
+    onSetup: selectionTargets.onSetupTableWithCaption
+  });
 };
 
 const addToolbars = (editor: Editor) => {